pyproject.toml
[project]
name = "sentinel-coach"
version = "0.3.0"
description = "Sentinel Coach — AI-powered fitness intelligence using TrainingPeaks + Garmin data"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "anthropic>=0.40.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.20.0",
    "cryptography>=43.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
fitness-bot = "fitness_ai_bot.main:main"
fitness-bot-api = "fitness_ai_bot.http_api:main"
